requirements.txt
urllib3
python-dotenv
requests
allure-python-commons
pytest
selene>=2.0.0b16
allure-pytest
selenium